remove ts, extraneous pages
