--- app/(tabs)/Notifications.tsx
@@ -1,34 +1,0 @@
-import { StyleSheet } from "react-native";
-import { Text, View } from "../../components/Themed";
-import EditScreenInfo from "../../components/EditScreenInfo";
-
-export default function Notifications() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Notifications</Text>
-      <View
-        style={styles.separator}
-        lightColor="#eee"
-        darkColor="rgba(255,255,255,0.1)"
-      />
-      <EditScreenInfo path="app/(tabs)/Notifications.tsx" />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  title: {
-    fontSize: 20,
-    fontWeight: "bold",
-  },
-  separator: {
-    marginVertical: 30,
-    height: 1,
-    width: "80%",
-  },
-});
