+++ app/_layout.jsx
@@ -55,7 +55,6 @@ function RootLayoutNav() {
           >
             <Stack.Screen name="(tabs)" />
             <Stack.Screen name="(auth)" />
-            <Stack.Screen name="modal" options={{ presentation: "modal" }} />
           </Stack>
         </Provider>
       </ThemeProvider>
