--- components/Themed.tsx
@@ -1,111 +1,0 @@
-/**
- * Learn more about Light and Dark modes:
- * https://docs.expo.io/guides/color-schemes/
- */
-
-import {
-  Text as DefaultText,
-  useColorScheme,
-  View as DefaultView,
-  PressableProps as DefaultPressableProps,
-  Pressable as DefaultPressable,
-  TextInput as DefaultTextInput,
-  ViewStyle,
-} from "react-native";
-
-import Colors from "../constants/Colors";
-
-export function useThemeColor(
-  props: { light?: string; dark?: string },
-  colorName: keyof typeof Colors.light & keyof typeof Colors.dark
-) {
-  const theme = useColorScheme() ?? "light";
-  const colorFromProps = props[theme];
-
-  if (colorFromProps) {
-    return colorFromProps;
-  } else {
-    return Colors[theme][colorName];
-  }
-}
-
-type ThemeProps = {
-  lightColor?: string;
-  darkColor?: string;
-};
-
-export type TextProps = ThemeProps & DefaultText["props"];
-export type ViewProps = ThemeProps & DefaultView["props"];
-export type PressableProps = ThemeProps & DefaultPressableProps;
-export type TextInputProps = ThemeProps & DefaultTextInput["props"];
-
-export function Text(props: TextProps) {
-  const { style, lightColor, darkColor, ...otherProps } = props;
-  const color = useThemeColor({ light: lightColor, dark: darkColor }, "text");
-
-  return (
-    <DefaultText
-      style={[{ color, fontFamily: "Karma" }, style]}
-      {...otherProps}
-    />
-  );
-}
-
-export function View(props: ViewProps) {
-  const { style, lightColor, darkColor, ...otherProps } = props;
-  const backgroundColor = useThemeColor(
-    { light: lightColor, dark: darkColor },
-    "background"
-  );
-
-  return <DefaultView style={[{ backgroundColor }, style]} {...otherProps} />;
-}
-
-export function Pressable(props: PressableProps) {
-  const { style, lightColor, darkColor, ...otherProps } = props;
-
-  const backgroundColor = useThemeColor(
-    { light: lightColor, dark: darkColor },
-    "buttonBackground"
-  );
-
-  const borderColor = useThemeColor(
-    { light: lightColor, dark: darkColor },
-    "borderColor"
-  );
-
-  return (
-    <DefaultPressable
-      style={[{ backgroundColor, borderColor }, style as ViewStyle]}
-      {...otherProps}
-    />
-  );
-}
-
-export function TextInput(props: TextInputProps) {
-  const { style, lightColor, darkColor, ...otherProps } = props;
-
-  const backgroundColor = useThemeColor(
-    { light: lightColor, dark: darkColor },
-    "textBoxBackground"
-  );
-
-  const textColor = useThemeColor({}, "text");
-
-  return (
-    <DefaultTextInput
-      placeholderTextColor={textColor + "8"}
-      style={[
-        {
-          backgroundColor,
-          padding: 8,
-          borderRadius: 8,
-          fontSize: 16,
-          fontFamily: "Karma",
-        },
-        style,
-      ]}
-      {...otherProps}
-    />
-  );
-}
