--- tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "expo/tsconfig.base",
-  "compilerOptions": {
-    "strict": true,
-    "skipLibCheck": true
-  }
-}
